IKASAN-1894 - Change jms and common maven artifacts to contain "spring", change package names to include spring
