+++ ikasaneip/component-factory/spring/common/src/main/java/org/ikasan/component/factory/spring/common/ConfigurationHandler.java
@@ -1,0 +1,88 @@
+/*
+ *  ====================================================================
+ *  Ikasan Enterprise Integration Platform
+ *
+ *  Distributed under the Modified BSD License.
+ *  Copyright notice: The copyright for this software and a full listing
+ *  of individual contributors are as shown in the packaged copyright.txt
+ *  file.
+ *
+ *  All rights reserved.
+ *
+ *  Redistribution and use in source and binary forms, with or without
+ *  modification, are permitted provided that the following conditions are met:
+ *
+ *   - Redistributions of source code must retain the above copyright notice,
+ *     this list of conditions and the following disclaimer.
+ *
+ *   - Redistributions in binary form must reproduce the above copyright notice,
+ *     this list of conditions and the following disclaimer in the documentation
+ *     and/or other materials provided with the distribution.
+ *
+ *   - Neither the name of the ORGANIZATION nor the names of its contributors may
+ *     be used to endorse or promote products derived from this software without
+ *     specific prior written permission.
+ *
+ *  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ *  AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ *  IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ *  DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+ *  FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ *  DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ *  SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ *  CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+ *  OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
+ *  USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ *  ====================================================================
+ *
+ */
+
+package org.ikasan.component.factory.spring.common;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.core.env.Environment;
+
+import java.lang.reflect.InvocationTargetException;
+
+public class ConfigurationHandler<C> {
+
+    private final Environment env;
+
+    public ConfigurationHandler(Environment env) {
+        this.env = env;
+    }
+
+    protected C configuration(String configPrefix, Class<C> clazz)
+    {
+        return BindConfigurationHelper.createConfigWithPrefixAndClassName(configPrefix, clazz, env);
+    }
+
+    private static Logger logger = LoggerFactory.getLogger(NonConfiguredResourceBaseComponentFactory.class);
+
+    protected C configuration(String configPrefix, String sharedConfigPrefix, Class<C> clazz)
+    {
+        C configuration = null;
+        try {
+            configuration = BindConfigurationHelper.createConfigWithPrefixAndClassName(configPrefix, clazz, env);
+        } catch (IkasanComponentFactoryException icfe){
+            logger.warn("Could not find properties with config prefix [{}] assuming all properties are shared with " +
+                    "prefix [{}]",configPrefix, sharedConfigPrefix);
+        }
+        C sharedConfiguration = BindConfigurationHelper.createConfigWithPrefix(clazz,env, sharedConfigPrefix);
+        try {
+            if (configuration == null){
+                configuration = clazz.getDeclaredConstructor().newInstance();
+            }
+            BeanMergeUtil.mergeSourceIntoTargetBean(configuration,sharedConfiguration);
+        } catch (InvocationTargetException | IllegalAccessException | NoSuchMethodException | InstantiationException e) {
+            throw new
+                    IkasanComponentFactoryException("Unable to merge properties with prefix [" + configPrefix
+                    + "] and sharedConfigPrefix [" + sharedConfigPrefix
+                    + "]  for configuration "
+                    + " of type " + clazz.getSimpleName());
+        }
+        configuration = null;
+        return sharedConfiguration;
+    }
+}
