+++ ikasaneip/component-factory/spring/jms/src/main/java/org/ikasan/component/factory/spring/jms/JmsComponentFactoryAutoConfiguration.java
@@ -37,50 +37,24 @@
  *
  */
 
-package org.ikasan.component.factory.common;
+package org.ikasan.component.factory.spring.jms;
 
 
-import org.apache.commons.collections.CollectionUtils;
-import org.apache.commons.collections.MapUtils;
-import org.springframework.beans.BeanUtils;
-import org.springframework.beans.BeanWrapper;
-import org.springframework.beans.BeanWrapperImpl;
+import org.ikasan.component.factory.spring.jms.consumer.JmsConsumerComponentFactory;
+import org.ikasan.component.factory.spring.jms.producer.JmsProducerComponentFactory;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
 
-import java.lang.reflect.InvocationTargetException;
-import java.util.Collection;
-import java.util.HashSet;
-import java.util.Map;
-import java.util.Set;
+@Configuration
+public class JmsComponentFactoryAutoConfiguration {
 
-/**
- * Util class to Merge One Beans Properties into another. Used by shared properties.
- */
-public class BeanMergeUtil {
-
-
-    public static <T> T mergeSourceIntoTargetBean(T source, T target) throws InvocationTargetException, IllegalAccessException {
-        BeanUtils.copyProperties(source, target, getNullPropertyNames(source));
-        return target;
+    @Bean
+    public JmsConsumerComponentFactory jmsConsumerComponentFactory(){
+        return new JmsConsumerComponentFactory();
     }
 
-    public static String[] getNullPropertyNames (Object source) {
-        final BeanWrapper src = new BeanWrapperImpl(source);
-        java.beans.PropertyDescriptor[] pds = src.getPropertyDescriptors();
-
-        Set<String> emptyNames = new HashSet<String>();
-        for(java.beans.PropertyDescriptor pd : pds) {
-            Object srcValue = src.getPropertyValue(pd.getName());
-            if (srcValue == null) emptyNames.add(pd.getName());
-            else if (srcValue instanceof Collection && CollectionUtils.isEmpty((Collection)srcValue)){
-                emptyNames.add(pd.getName());
-            }
-            else if (srcValue instanceof Map && MapUtils.isEmpty((Map)srcValue)){
-                emptyNames.add(pd.getName());
-            }
-
-        }
-
-        String[] result = new String[emptyNames.size()];
-        return emptyNames.toArray(result);
+    @Bean
+    public JmsProducerComponentFactory jmsProducerComponentFactory(){
+        return new JmsProducerComponentFactory();
     }
 }
