+++ ikasaneip/component-factory/spring/jms/src/main/resources/META-INF/spring.factories
@@ -1,1 +1,1 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=org.ikasan.component.factory.jms.JmsComponentFactoryAutoConfiguration
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=org.ikasan.component.factory.spring.jms.JmsComponentFactoryAutoConfiguration
